@@ -1,4 +1,4 @@
-// swift-tools-version: 5.7
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
     name: "fyno",
     platforms: [
-        .iOS(.v12)  // Specify your minimum target iOS version
+        .iOS(.v13)  // Specify your minimum target iOS version
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
