@@ -1,4 +1,4 @@
-// Generated by Apple Swift version 6.0 effective-5.10 (swiftlang-6.0.0.9.10 clang-1600.0.26.2)
+// Generated by Apple Swift version 6.0.3 (swiftlang-6.0.3.1.10 clang-1600.0.30.1)
 #ifndef FYNO_SWIFT_H
 #define FYNO_SWIFT_H
 #pragma clang diagnostic push
